feat(auth): clerk auth

+++ packages/backend/convex/users.ts
@@ -3,6 +3,10 @@ import { mutation, query } from "./_generated/server";
 export const getmany = query({
   args: {},
   handler: async (ctx) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if(identity == null){
+      throw new Error("unauthorized")
+    }
     const users = await ctx.db.query("users").collect();
     return users;
   },
@@ -12,6 +16,10 @@ export const getmany = query({
 export const addUser = mutation({
   args: {},
   handler: async(ctx) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if(identity == null){
+      throw new Error("unauthorized")
+    }
     const userId = await ctx.db.insert("users", {
       name: "Abhishek gusain"
     })
@@ -21,6 +29,10 @@ export const addUser = mutation({
 export const removeUsers = mutation({
   args: {},
   handler: async(ctx) => {
+    const identity = await ctx.auth.getUserIdentity();
+    if(identity == null){
+      throw new Error("unauthorized")
+    }
     const allUsers = await ctx.db.query("users").collect();
     allUsers.map(async(user) => {
       await ctx.db.delete(user._id)
